+++ controllers/src/main/java/com.leverxblog/controllers/CommentController.java
@@ -1,16 +1,17 @@
 package com.leverxblog.controllers;
 
+import com.leverxblog.dto.ArticleDto;
 import com.leverxblog.dto.CommentDto;
-import com.leverxblog.dto.UserDto;
+import com.leverxblog.services.implementation.ArticleService;
 import com.leverxblog.services.implementation.CommentService;
+import com.leverxblog.services.implementation.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.*;
 
-import javax.swing.*;
-import javax.xml.ws.Response;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -20,15 +21,26 @@ import java.util.UUID;
 @RequestMapping("/articles")
 public class CommentController {
     private CommentService commentService;
+    private UserService userService;
+    private ArticleService articleService;
 
     @Autowired
-    public CommentController(CommentService commentService) {
+    public CommentController(CommentService commentService, UserService userService, ArticleService articleService) {
         this.commentService = commentService;
+        this.userService = userService;
+        this.articleService = articleService;
     }
 
     @PostMapping("/{articleId}/comments")
-    public ResponseEntity<Object> addNewComment(@PathVariable UUID articleId, @RequestBody CommentDto commentDto) {
+    public ResponseEntity<Object> addNewComment(@PathVariable UUID articleId, @RequestBody CommentDto commentDto,
+                                                Authentication authentication) {
+
+        String login= authentication.getName();
+        UUID userId = userService.getByLogin(login).getId();
+
+        commentDto.setUserId(userId);
         commentDto.setArticleId(articleId);
+
         Map id = Collections.singletonMap("id", commentService.add(commentDto));
         return new ResponseEntity<>(id, HttpStatus.CREATED);
     }
@@ -54,9 +66,23 @@ public class CommentController {
     }
 
     @DeleteMapping("/{articleId}/comments/{commentId}")
-    public ResponseEntity<Void> deleteComment(@PathVariable UUID articleId, @PathVariable UUID commentId) throws Exception{
-        //to add checking authorship
-        commentService.delete(commentId);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<Void> deleteComment(@PathVariable UUID articleId, @PathVariable UUID commentId,
+                                              Authentication authentication) throws Exception{
+
+        CommentDto comment = commentService.getById(commentId);
+
+        if (comment == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        ArticleDto article = articleService.getById(comment.getArticleId());
+        String login= authentication.getName();
+        UUID userId = userService.getByLogin(login).getId();
+
+        if ((comment.getUserId()).compareTo(userId)==0 || (article.getUserEntity_id()).compareTo(userId)==0){
+            commentService.delete(commentId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 }
